+++ tools/Az.Tools.Installer/ChangeLog.md
@@ -1,4 +1,5 @@
 ## Upcoming Release
+* Fix Update-Module errors when Az modules are installed using Az.Tools.Installer [#21262]
 
 ## 0.2.2
 * Updated the docs of parameter `Scope` [#16768]